fix(seo): apply route metadata and indexation signals
