+++ src/app/projects/metadata.ts
@@ -3,7 +3,7 @@ import type { Metadata } from "next";
 const BASE_URL = "https://www.dvillagrans.dev";
 
 export const metadata: Metadata = {
-  title: "Projects Archive — Diego Villagran",
+  title: "Projects Archive",
   description: "Complete archive of data science, analytics, ML engineering, and full-stack projects by Diego Villagran. 15+ systems built across AI, ETL, and web platforms.",
   openGraph: {
     title: "Projects Archive | Diego Villagran",
