+++ src/app/projects/nyc/page-client.tsx
@@ -1,0 +1,28 @@
+"use client";
+
+import { useLanguage } from "@/i18n/LanguageContext";
+import { nycEn, nycEs } from "@/i18n/dictionaries/nyc";
+import { EditorialCaseStudy } from "@/components/case-study/EditorialCaseStudy";
+import { CaseStudyExternalActions } from "@/components/case-study/CaseStudyExternalActions";
+import { NycProjectSchema } from "@/components/ui/SchemaOrg";
+
+export default function NycCaseStudyPage() {
+  const { language } = useLanguage();
+  const dict = language === "es" ? nycEs : nycEn;
+
+  return (
+    <EditorialCaseStudy
+      slug="nyc"
+      dict={dict}
+      schema={<NycProjectSchema />}
+      actions={
+        <CaseStudyExternalActions
+          dashboardHref={dict.linkDashboard}
+          dashboardLabel={dict.links.dashboard}
+          repoHref={dict.linkRepo}
+          repoLabel={dict.links.repo}
+        />
+      }
+    />
+  );
+}
