+++ src/app/robots.ts
@@ -1,5 +1,7 @@
 import type { MetadataRoute } from 'next';
 
+const BASE_URL = 'https://www.dvillagrans.dev';
+
 export default function robots(): MetadataRoute.Robots {
   return {
     rules: {
@@ -7,6 +9,6 @@ export default function robots(): MetadataRoute.Robots {
       allow: '/',
       disallow: ['/api/', '/_next/'],
     },
-    sitemap: 'https://www.dvillagrans.dev/sitemap.xml',
+    sitemap: `${BASE_URL}/sitemap.xml`,
   };
 }
